+++ zigzag.legendweathers/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Majora Mask invocation now all the same player
 - Possible fix for Majora fog not being enabled at start of moon
 - Majora compatible with Imperium time manipulation
+- Majora music will not play if volume in config is specified 0
 
 ## 1.1.11
 - Added custom colors for some combined weathers names